+++ android/app/src/main/java/com/clipsync/android/MainActivity.kt
@@ -590,6 +590,7 @@ private fun ClipSyncApp(
                         onAutoExpireChange = preferencesViewModel::setAutoExpire,
                         onBootRestoreChange = preferencesViewModel::setBootRestore,
                         onImageSyncChange = preferencesViewModel::setImageSync,
+                        onAutoApplyImagesChange = preferencesViewModel::setAutoApplyImages,
                         pairedDeviceName = healthState.pairedPeerName,
                         onOpenConduit = {
                             pairingOpen = false